Extract runtime_test_utils
